drop fallback codes for Go 1.7
