+++ tracer_register.go
@@ -1,5 +1,3 @@
-// +build go1.5
-
 package proxy
 
 import (
